[PATCH] parisc: make atomic_t 32-bit
Make atomic_t an int instead of a long to match the other architectures

Committed-by: Matthew Wilcox <willy@parisc-linux.org>

+++ include/asm-parisc/atomic.h
@@ -52,7 +52,7 @@ extern spinlock_t __atomic_hash[ATOMIC_HASH_SIZE] __lock_aligned;
  * Cache-line alignment would conflict with, for example, linux/module.h
  */
 
-typedef struct { volatile long counter; } atomic_t;
+typedef struct { volatile int counter; } atomic_t;
 
 
 /* This should get optimized out since it's never called.
